koreai/src/main/java/com/github/deutschebank/symphony/koreai/KoreaResponse.java
package com.github.deutschebank.symphony.koreai;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;

import java.util.List;

/**
 * @author rodriva
 */
@JsonIgnoreProperties(ignoreUnknown = true)
public class KoreaResponse {
    private String text;
    @JsonProperty(value = "isTemplate")
    private boolean template;
    private List<Errors> errors;

    public String getText() {
        return text;
    }

    public void setText(String text) {
        this.text = text;
    }

    public boolean isTemplate() {
        return template;
    }

    public void setTemplate(boolean template) {
        this.template = template;
    }

    public List<Errors> getErrors() {
        return errors;
    }

    public void setErrors(List<Errors> errors) {
        this.errors = errors;
    }

    static class Errors {
        private String msg;
        private int code;

        public String getMsg() {
            return msg;
        }

        public void setMsg(String msg) {
            this.msg = msg;
        }

        public int getCode() {
            return code;
        }

        public void setCode(int code) {
            this.code = code;
        }
    }
}
